refactor(app): rename ResourceConfiguration to MonitoringConfiguration
