+++ centreon/config.new/packages/api_platform.yaml
@@ -18,7 +18,7 @@ api_platform:
 
     mapping:
         paths:
-            - '%kernel.project_dir%/src/App/ResourceConfiguration/Infrastructure/ApiPlatform/Resource/'
+            - '%kernel.project_dir%/src/App/MonitoringConfiguration/Infrastructure/ApiPlatform/Resource/'
 
     swagger:
          api_keys:
