+++ centreon/tests/php/App/ActivityLogging/Domain/Event/LogActivityEventHandlerTest.php
@@ -27,10 +27,10 @@ use App\ActivityLogging\Domain\Aggregate\ActionEnum;
 use App\ActivityLogging\Domain\Aggregate\ActivityLog;
 use App\ActivityLogging\Domain\Aggregate\TargetTypeEnum;
 use App\ActivityLogging\Domain\Event\LogActivityEventHandler;
-use App\ResourceConfiguration\Domain\Aggregate\ServiceCategory;
-use App\ResourceConfiguration\Domain\Aggregate\ServiceCategoryId;
-use App\ResourceConfiguration\Domain\Aggregate\ServiceCategoryName;
-use App\ResourceConfiguration\Domain\Event\ServiceCategoryCreated;
+use App\MonitoringConfiguration\Domain\Aggregate\ServiceCategory;
+use App\MonitoringConfiguration\Domain\Aggregate\ServiceCategoryId;
+use App\MonitoringConfiguration\Domain\Aggregate\ServiceCategoryName;
+use App\MonitoringConfiguration\Domain\Event\ServiceCategoryCreated;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Tests\App\ActivityLogging\Infrastructure\Double\FakeActivityLogFactory;
